Rearranged Code
*Moved code for commands to IHA_Command.py
*Fixed IHA comparing command to intent
